물품 체크리스트 카테고리 조회 완료(close #6)

+++ app/src/main/java/com/example/graduationproject/Api/Response/CategoryCheckListResponse.kt
@@ -12,6 +12,6 @@ data class CategoryCheckListResponse (
     @SerializedName("productName")
     val productName:String,
     @SerializedName("productPrice")
-    val productPrice:Int
+    val productPrice:Long
 
         )
